+++ src/libopensc/pkcs15-pubkey.c
@@ -76,8 +76,8 @@ int sc_pkcs15_decode_pukdf_entry(struct sc_pkcs15_card *p15card,
         struct sc_context *ctx = p15card->card->ctx;
         struct sc_pkcs15_pubkey_info info;
 	int r;
-	int usage_len = sizeof(info.usage);
-	int af_len = sizeof(info.access_flags);
+	size_t usage_len = sizeof(info.usage);
+	size_t af_len = sizeof(info.access_flags);
 	struct sc_asn1_entry asn1_com_key_attr[6], asn1_com_pubkey_attr[1];
 	struct sc_asn1_entry asn1_rsakey_attr[4], asn1_rsa_type_attr[2];
 	struct sc_asn1_entry asn1_dsakey_attr[2], asn1_dsa_type_attr[2];
@@ -151,7 +151,7 @@ int sc_pkcs15_encode_pukdf_entry(struct sc_context *ctx,
 						    asn1_com_key_attr,
 						    asn1_com_pubkey_attr, asn1_dsa_type_attr };
 	int r;
-	int af_len, usage_len;
+	size_t af_len, usage_len;
 
         sc_copy_asn1_entry(c_asn1_pubkey, asn1_pubkey);
         sc_copy_asn1_entry(c_asn1_rsa_type_attr, asn1_rsa_type_attr);
